Updates to latest maven and plugins (#3606)
This updates to the latest maven 3.9.6 and removes the takari wrapper
plugin, as that's been obviated upstream for three years. This also
bumps all the maven plugins used, which is good as many of them were
pre-release versions formerly.

This intentionally only does maven stuff in one PR. There is a lot of
out-of-date deps that really need to be addressed before a patch of the
current codebase before doing anything else. This is just one step. Next
will include test and main deps as well docker and github actions stuff,
in any order.

Signed-off-by: Adrian Cole <adrian@tetrate.io>