Sets supported JDK range explicitly to 11-15
Before, we were using an implicit version of the JDK when releasing, had
the pom.xml limited to a non-LTS version (15).

This sets the same up explicitly in CI, and corrects CI to use latest
action versions and ubuntu. A later change should update the ceiling
build version to an LTS, at least 17 if not 21. Currently, we can't as
there is a compilation failure. I'll look into it personally in the next
PR.

Signed-off-by: Adrian Cole <adrian@tetrate.io>