Fixes flakey ZipkinExtensionTest
This attempts to fix, rather than ignore (which was done previously and
I removed in the last commit) a flakey test. It appears likely that not
using resource management at all in this test could result in flakes.
Even if not, it is worth removing doubt around this.

```
 Error:  Tests run: 10, Failures: 0, Errors: 1, Skipped: 0, Time elapsed: 1.739 s <<< FAILURE! -- in zipkin2.junit5.ZipkinExtensionTest
Error:  zipkin2.junit5.ZipkinExtensionTest.postSpans_disconnectDuringBody -- Time elapsed: 0.054 s <<< ERROR!
java.io.IOException: unexpected end of stream on http://localhost:53245/...
	at okhttp3.internal.http1.Http1ExchangeCodec.readResponseHeaders(Http1ExchangeCodec.kt:209)
--snip--
 Caused by: java.io.EOFException: \n not found: limit=0 content=…
	at okio.RealBufferedSource.readUtf8LineStrict(RealBufferedSource.kt:332)
	at okhttp3.internal.http1.HeadersReader.readLine(HeadersReader.kt:29)
	at okhttp3.internal.http1.Http1ExchangeCodec.readResponseHeaders(Http1ExchangeCodec.kt:179)
	... 84 more
```

Signed-off-by: Adrian Cole <adrian@tetrate.io>