docker: adds zipkin-activemq image to remove junit 4 rule dependency
This replaces the in-VM approach for testing activemq which has been
flakey and also results in a JUnit 4 rule dependency.

Note: This uses the JDK image, not the JRE, because ActiveMQ relies on
JMX even when disabled via configuration. This is a better choice than
making our production JRE image include it.

Once this is deployed I can make zipkin-reporter use it per below:
 https://github.com/openzipkin/zipkin-reporter-java/pull/220

I can also change our collector tests here to use it instead of the
JUnit 4 rule.

Signed-off-by: Adrian Cole <adrian@tetrate.io>