docker: uses health check in examples and documents dependencies job
This uses health check in the docker-compose examples to reduce the
amount of warnings end users see.

It also fixes a minor issue on the dependencies job, as well adds
instructions for how to run it manually. This is needed for folks who
don't want to wait an hour to see the graph.

Fixes #3529

Signed-off-by: Adrian Cole <adrian@tetrate.io>