docker: updates Elasticsearch 7.x image to Elastic licensed 7.17.16
Before, we rev-locked our Elasticsearch 7.x image to the last OSS
licensed version 7.10.2. This changes to the Elastic licensed current
version, 7.17.16, as that's what people would run in production.

Note: As this is a test image, not a production image, usage should not
violate the Elastic Basic license developers typically use, and is the
default.

Signed-off-by: Adrian Cole <adrian@tetrate.io>