Makes all except core jar Java 11+
This moves all code to minimum Java 11, except the core jar, which
remains (mostly) Java 1.6 for use in zipkin-reporter.

Notably, this pushes animal-sniffer to the core module, as the others
can rely on the release property of modern compilers.

Java 11 instead of 11 or 21 because that's the last SDK that can also
compile 1.6.

Note: this bumps to the next minor as it is worth communicating.

Signed-off-by: Adrian Cole <adrian@tetrate.io>