Prepares for 3.0 by updating to Spring Boot 3 and floor Java 8 (#3684)
We currently still have a problem as the latest version shows up as 3.0.0-rc0 due to an accidental release. Moreover, a request to delete that version hasn't happened. This switches to 3.x by dropping support for Java 6 and moving to Spring Boot 3.

As noted, recent Zipkin Reporter 3 and Brave 6 decouple instrumentation from this core jar. While it is still in use for other formats, the call sites using them are all minimum Java 8 anyway. By updating to floor Java 8, we can use later JDKs to release than 11, though this change doesn't include doing that, yet.

As Spring Boot requires JRE 17, we update the floor JDK to 17. However, zipkin-dependencies is currently stuck at 11, so until that's not the case, this leaves collector and storage modules at Java 11, even if the server can only run in 17+.

Signed-off-by: Adrian Cole <adrian@tetrate.io>