ci: adds lint check and fixed bad links
We should quote any example links and only add links for things that are
always up, or will be up due following instructions.

For example, we should add exceptions for localhost services that run in
result of test commands for zipkin or the UI.

See https://github.com/openzipkin/zipkin-gcp/pull/212 for the first impl
of this.

Signed-off-by: Adrian Cole <adrian@tetrate.io>