cassandra: cleans up code and tests
This cleans up deprecation and style issues found in IntelliJ analysis.
It also migrates docker tests to junit-jupiter and fixes the logging
configuration which broke when we updated to SLF4J 2.x.

This also updates all use of docker test images to the latest, in
attempts to share base image layers.

Signed-off-by: Adrian Cole <adrian@tetrate.io>