Updates to Armeria 1.27.0 and fixes Eureka VIP registration bug
This updates to latest Armeria 1.27.0, which removes some tech debt.
This also fixes a bug in the default behavior of Armeria, which is that
it sets a `host:port` value for the VIP address in Eureka. While the
`instanceId` needs to be qualified with port, the VIP address must not,
as there's already a port field.

Signed-off-by: Adrian Cole <adrian@tetrate.io>