lens: fixes BASE_PATH handling and re-aligns to old resource paths
The last commit moved things forward, but it broke how we used the
BASE_PATH variable. This reverts to the old logic and fixes up the
build so it behaved like the last one.

Along the way, I discovered some tests were disabled, and as able to
fix them except one that I need @anuraaga or similar's help on.

@SamTV12345 as a matter of some urgency, once this PR is merged, can you
follow-up and test i18n? I noticed the language toggle doesn't work, as
it just is always english.

Signed-off-by: Adrian Cole <adrian@tetrate.io>